# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The ETSI MEC ISG MEC013 Location API described using OpenAPI",
        "title": "ETSI GS MEC 013 - Location API",
        "contact": {
            "name": "Iulisloi Zacarias",
            "url": "https://github.com/izacarias"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "3.1.1"
    },
    "host": "localhost:8080",
    "basePath": "/location/v3/",
    "paths": {
...
        "/queries/users": {
            "get": {
                "description": "The GET method is used to query location information about a specific UE or a group of Ues.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "Query location information about a specific UE or a group of UEs",
                "operationId": "usersGET",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Zone ID",
                        "name": "zoneId",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Access Point ID",
                        "name": "accessPointId",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "User address",
                        "name": "address",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.UserInfoList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "User Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/queries/zones": {
            "get": {
                "description": "The GET method is used to query the information about one or more specific zones or a list of zones.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "Retrieves information about one or more specific zones or a list of zones.",
                "operationId": "zonesGET",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Zone ID",
                        "name": "zoneId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ZoneList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/queries/zones/{zoneId}": {
            "get": {
                "description": "The GET method is used to query the information about a specific zone.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "Query information about a specific zone",
                "operationId": "zoneGetById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Zone ID",
                        "name": "zoneId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ZoneInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "responses.LocationInfo": {
            "type": "object",
            "properties": {
                "altitude": {
                    "description": "Location altitude relative to the WGS84 ellipsoid surface.",
                    "type": "number"
                },
                "latitude": {
                    "description": "Location latitude, expressed in the range -90° to +90°. Cardinality greater than one only if \\\"shape\\\" equals 7.",
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "longitude": {
                    "description": "Location longitude, expressed in the range -180° to +180°. Cardinality greater than one only if \\\"shape\\\" equals 7.",
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "shape": {
                    "description": "Shape information, as detailed in [14], associated with the reported location coordinate:\n\t1 = Ellipsoid_Arc\n\t2 = ellipsoid_Point\n\t3 = ellipsoid_Point_Altitude\n\t4 = ellipsoid_Point_Altitude_Uncert_Ellipsoid\n\t5 = ellipsoid_Point_Uncert_Circle\n\t6 = ellipsoid_Point_Uncert_Ellipse\n\t7 = polygon",
                    "allOf": [
                        {
                            "$ref": "#/definitions/responses.LocationInfoShape"
                        }
                    ]
                }
            }
        },
        "responses.LocationInfoShape": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5,
                6,
                7
            ],
            "x-enum-varnames": [
                "LocationInfoShapeN1",
                "LocationInfoShapeN2",
                "LocationInfoShapeN3",
                "LocationInfoShapeN4",
                "LocationInfoShapeN5",
                "LocationInfoShapeN6",
                "LocationInfoShapeN7"
            ]
        },
...
        "responses.ProblemDetails": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Type of the problem, typically a URI that identifies the problem type",
                    "type": "string"
                },
                "status": {
                    "description": "Instance string `json:\"instance,omitempty\"`",
                    "type": "integer"
                }
            }
        },
...
        "responses.TimeStamp": {
            "type": "object",
            "properties": {
                "nanoSeconds": {
                    "description": "The nanoseconds part of the time.",
                    "type": "integer"
                },
                "seconds": {
                    "description": "The seconds part of the time. Time is defined as Unix-time since January 1, 1970, 00:00:00 UTC.",
                    "type": "integer"
                }
            }
        },
        "responses.UserInfo": {
            "type": "object",
            "properties": {
                "AccessPointId": {
                    "description": "AccessPointId The identity of the access point the user is currently on, see note 1.",
                    "type": "string"
                },
                "address": {
                    "description": "Address Address of user (e.g. 'sip' URI, 'tel' URI, 'acr' URI) currently on the access point, see note 1.",
                    "type": "string"
                },
                "ancillaryInfo": {
                    "description": "AncillaryInfo Reserved for future use.",
                    "type": "string"
                },
                "locationInfo": {
                    "description": "Location of the User",
                    "allOf": [
                        {
                            "$ref": "#/definitions/responses.LocationInfo"
                        }
                    ]
                },
                "resourceURL": {
                    "description": "ResourceURL Self-referring URL, see note 1.",
                    "type": "string"
                },
                "timestamp": {
                    "description": "Date and time that location was collected.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/responses.TimeStamp"
                        }
                    ]
                },
                "zoneId": {
                    "description": "ZoneId The identity of the zone the user is currently within, see note 1.",
                    "type": "string"
                }
            }
        },
        "responses.UserInfoList": {
            "type": "object",
            "properties": {
                "resourceURL": {
                    "description": "ResourceURL Self-referring URL, see note 1.",
                    "type": "string"
                },
                "user": {
                    "description": "User List of users, see note 1.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.UserInfo"
                    }
                }
            }
        },
        "responses.ZoneInfo": {
            "type": "object",
            "properties": {
                "numberOfAccessPoints": {
                    "description": "The number of access points within the zone",
                    "type": "integer"
                },
                "numberOfUnserviceableAccessPoints": {
                    "description": "Number of inoperable access points within the zone.",
                    "type": "integer"
                },
                "numberOfUsers": {
                    "description": "The number of users currently on the access point.",
                    "type": "integer"
                },
                "resourceURL": {
                    "description": "Self referring URL",
                    "type": "string"
                },
                "zoneId": {
                    "description": "Identifier of zone",
                    "type": "string"
                }
            }
        },
        "responses.ZoneInfoList": {
            "type": "object",
            "properties": {
                "resourceURL": {
                    "type": "string"
                },
                "zone": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.ZoneInfo"
                    }
                }
            }
        },
        "responses.ZoneList": {
            "type": "object",
            "properties": {
                "zoneList": {
                    "$ref": "#/definitions/responses.ZoneInfoList"
                }
            }
        }
    }
}